This is a purely mechanical movement of the majority of `DataStore`'s methods into separate module files in an attempt to speed up compilation (particularly incremental compilation). On my machine, it looks like a modest success, cutting incremental compilation down by ~7 seconds for both test and non-test builds:

| Operation | main (`2a2a7f7e`) | this branch (`647e336e`) |
| --- | --- | --- |
| clean build, lib+bin | 2m 55s | 2m 48s |
| incremental build, lib+bin, `datastore.rs` modified | 45s | 38s |
| clean build, tests | 5m 5s | 4m 49s |
| incremental build, tests, `datastore.rs` modified | 1m 14s | 1m 8s |

where "lib+bin" is `cargo build -p omicron-nexus` and "tests" is `cargo build -p omicron-nexus --tests` (i.e., the build of the tests without running them). Modifying files underneath `datastore/` on this branch results in identical timings to modifying `datastore/mod.rs` (which is surprising to me, but it's quite reliable).

These timings are all debug builds; if there's interest I can collect them for release builds also.