@david-crespo 2022-07-15
Pretty cool, didn't know this would have this effect. 10% on incremental builds is pretty good!