@jgallagher 2022-07-15
@smklein suggested it when we were chatting about ideas as a way to speed up incremental times since diesel is one of the biggest time sinks. I don't understand why this would speed up the clean builds, though!