@jgallagher 2022-07-16
Out of curiosity, what happens if you modify `datastore.rs` (e.g., add a word to the doc comment at the top of the file) instead of just `touch`'ing it?